requirements.txt
mysql-connector
numpy
opencv-python